Extracted Classes MemberAnalyzer and TypeAnalyzer
